backend/requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
apache-beam==2.49.0
array-record==0.4.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
avro-python3==1.10.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.1.0
crcmod==1.7
cycler==0.11.0
Cython==3.0.0
dill==0.3.1.1
dm-tree==0.1.8
dnspython==2.4.1
docopt==0.6.2
etils==1.4.0
fastavro==1.8.2
fasteners==0.18
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.41.1
frozenlist==1.4.0
gast==0.4.0
gin-config==0.5.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
h5py==3.9.0
hdfs==2.7.0
httplib2==0.22.0
idna==3.4
immutabledict==3.0.0
importlib-resources==6.0.0
itsdangerous==2.1.2
jax==0.4.14
Jinja2==3.1.2
joblib==1.3.1
kaggle==1.5.16
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.6
lvis==0.5.3
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
ml-dtypes==0.2.0
multidict==6.0.4
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.6.1
openai==0.27.8
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
orjson==3.9.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
portalocker==2.7.0
promise==2.3
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pydot==1.4.2
pymongo==4.4.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==5.3.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sacrebleu==2.2.0
scikit-learn==1.3.0
scipy==1.11.1
sentencepiece==0.1.99
seqeval==1.2.2
simplejsondb==0.4.0
six==1.16.0
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.1
tensorflow-addons==0.21.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-hub==0.14.0
tensorflow-io==0.32.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
tensorflow-model-optimization==0.7.5
tensorflow-text==2.12.1
termcolor==2.3.0
text-unidecode==1.3
tf-models-official==2.12.1
tf-slim==1.1.0
threadpoolctl==3.2.0
toml==0.10.2
tqdm==4.65.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.14.1
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
